+++ dkb/dkb.py
@@ -5,14 +5,13 @@ Created on 22.06.2021
 @author: wakl8754
 '''
 from typing import Any
-import pandas as pd 
 import csv
 import hashlib
 # import user packages
-from fl.fileLoader import FileLoader
+from dkb.fl.fileLoader import FileLoader
+from dkb.cfg import ResponseCode as RC
 
-
-DEBUGLEVEL = 1
+DEBUGLEVEL = 3
 INFO = False
 
 class DKB(object):
@@ -20,50 +19,13 @@ class DKB(object):
     Performs parsing of the CSV file from DKB
     '''
     # the line here to find header line
-    __header_line = 6               
-    # headers to map
-    __header = [
-                'date',
-                'booking-date',
-                'text',
-                'debitor',
-                'verwendung',
-                'konto',
-                'blz',
-                'value',
-                'debitor-id',
-                'Mandats reference',
-                'Customer reference']
+                
+   
     def __init__(self):
         ''' @pth input pathlib.Path type see https://docs.python.org/3/library/pathlib.html 
             pth to the folder with csv files path shall of the type pathlib.Path
         '''        
         pass
-                
-    def parseDkbData(self):
-        pieces = []
-        df = None
-        
-        self.csv_files = self.loader.getCsvFilesList()
-        if len(self.csv_files) == 0:
-            if DEBUGLEVEL > 0: print('# list of files is empty #')
-            return None
-        for csv in self.csv_files:
-            df = self._getData(csv)
-            if df is not None:
-                pieces.append(df)
-        if len(pieces) == 1:
-            return df
-        elif len(pieces) > 1:     
-            if INFO: print('# parseDkbData: # Concatenate DataFrames with panda #')       
-            return pd.concat(pieces)
-        else:
-            if DEBUGLEVEL > 0: print('# parseDkbData: # list of pieces is empty #')
-            return df
-    
-    def getDF(self):
-        self.csv_files = self.loader.getCsvFilesList()
-        self._getData(self.csv_files[0])
     
     def get_meta(self, csv_file):
         dkb_format = self._checkDkbFormat(csv_file)
@@ -97,35 +59,6 @@ class DKB(object):
                         if DEBUGLEVEL > 1: print(f'# DKB #:  Kontostand: {self.metaDic["Kontostand"]} #')
                     elif nrLines > 6: return self.metaDic
                     nrLines +=1
-           
-    def _getData(self, csv_file):       
-        '''
-        reads the CSV file and creates data frame with the parsed data. Panda df is returned.
-        https://pandas.pydata.org/pandas-docs/stable/reference/frame.html
-        ''' 
-        if DEBUGLEVEL > 0: print(f"# DKB #: csv file: {csv_file} ")
-        dkb_format = self._checkDkbFormat(csv_file)
-        
-        try:
-            if dkb_format:
-                df = pd.read_csv(csv_file,
-                    skiprows = self.__header_line,                  
-                    #encoding='utf-8', # needs to be None to be able to read German text                  
-                    encoding_errors = "ignore", 
-                    #warn_bad_lines=True, depricated
-                    delimiter=";",
-                    skipinitialspace = True,
-                    parse_dates = [0,1], # to parse these colums as date
-                    infer_datetime_format=True
-                    )
-                
-                df.columns = self.__header
-                return df
-            else:
-                return None
-        except: 
-            print(f"# DKB #: File not found at: {csv_file}, or file is corrupted")
-            return None
         
     def _checkDkbFormat(self, csv_file) -> Any:   
         try:
@@ -133,11 +66,11 @@ class DKB(object):
                 reader = csv.reader(csvfile, delimiter=';')
                 ctr = 0
                 for row in reader:
-                    if DEBUGLEVEL > 3: print(', '.join(row))
-                    # TODO: check how many columns the csv file has
-                    if ctr > 20: break
+                    if DEBUGLEVEL > 4: print(', '.join(row))
+                    if ctr > 7: break
                     if "Kontonummer" in row:
                         if DEBUGLEVEL > 1: print('# DKB #:  DKB format csv #')
+                        konto = row[1]
                         return True                    
                     ctr =+ 1
                 return False
@@ -145,41 +78,34 @@ class DKB(object):
             print(f"# DKB #: File not found at: {csv_file}, or file is corrupted")
             return False
     
-    def get_dkb_df(self, csv_file):
+    def _calculate_checksum(self, data):
+        return hashlib.md5(','.join(data).encode('utf-8')).hexdigest()
+
+    def get_lines_as_list_of_dicts(self, csv_file):
+        lines = []
+        dkb_dict = {}
         if self._checkDkbFormat(csv_file):
-            # with open(hashlib.csv_file, 'rb') as f:
-            #     hash = hashlib.file_digest(f, 'sha256')
-            #     self._hach_table.append(hash)
-            #     print(f'{hash}')
-            df = pd.read_csv(csv_file,
-                    skiprows = self.__header_line,                  
-                    #encoding='utf-8', # needs to be None to be able to read German text                  
-                    encoding_errors = "ignore", 
-                    #warn_bad_lines=True, depricated
-                    delimiter=";",
-                    skipinitialspace = True,
-                    parse_dates = [0,1], # to parse these colums as date
-                    infer_datetime_format=True
-                    )
-                
-            df.columns = self.__header
-            return df
-        else:
-            return None
-    
-    def get_dkb_df_from_folder(self, pth):
-        pieces = []
-        df = None 
-        self.loader = FileLoader(pth)    
-        self.csv_files = self.loader.getCsvFilesList()
-        if len(self.csv_files) == 0:
-            if DEBUGLEVEL > 0: print('# list of files is empty #')
-            return None
-        for csv in self.csv_files:
-            df = self.get_dkb_df(csv)
-            if df is not None:
-                pieces.append(df)
-        if len(pieces) > 1:     
-            if INFO: print('# parseDkbData: # Concatenate DataFrames with panda #')       
-            df = pd.concat(pieces)
-        return df
+            with open(csv_file, newline='') as csvfile:
+                reader = csv.reader(csvfile, delimiter=';')
+                ctr = 0
+                for row in reader:
+                    ctr += 1
+                    if DEBUGLEVEL > 3: print(', '.join(row))
+                    if ctr < 8: continue
+                    dkb_dict = {}
+                    
+                    dkb_dict['date'] = row[0]
+                    dkb_dict['booking-date'] = row[1]
+                    dkb_dict['text'] = row[2]
+                    dkb_dict["debitor"] = row[3]
+                    dkb_dict['verwendung'] = row[4]
+                    dkb_dict['konto'] = row[5]
+                    dkb_dict['blz'] = row[6]
+                    dkb_dict['value'] = float(row[7].replace('.','').replace(',', '.'))
+                    dkb_dict['debitor-id'] = row[8]                    
+                    dkb_dict['mandats_ref'] = row[9]
+                    dkb_dict['customer_ref'] = row[10]
+                    dkb_dict['checksum'] = hashlib.md5(','.join(row).encode('utf-8')).hexdigest()
+                    lines.append(dkb_dict)                    
+            if DEBUGLEVEL > 1: print(f'get_lines_as_list_of_dicts: {len(lines)}')
+        return (RC.OK, lines)